+++ test/core/parse.test.js
@@ -23,4 +23,13 @@ suite(".parse", function() {
       Parsimmon.of("kaboom").parse();
     });
   });
+
+  context("The input is a buffer.", function() {
+    it("Formats errors correctly.", function() {
+      var parser = Parsimmon.Binary.byte(0);
+      assert.throws(function() {
+        parser.tryParse(Buffer.from([0xf]));
+      }, /at byte/);
+    });
+  });
 });